requirements.txt
tensorflow[and-cuda]==2.18
flask==3.1.3
onnx
tf2onnx
onnxruntime-gpu==1.19.0
tensorrt==10.0.1
